requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alabaster==0.7.12         # via sphinx
appdirs==1.4.4            # via black, virtualenv
asgiref==3.3.1            # via django
attrs==20.3.0             # via pytest, pytest-mypy
babel==2.9.0              # via sphinx
black==20.8b1             # via -r requirements.in, pytest-black
bleach==3.2.1             # via -r requirements.in
certifi==2020.11.8        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black, pip-tools
coverage==5.3             # via pytest-cov
distlib==0.3.1            # via virtualenv
django-yaa-settings==1.1.0  # via -r requirements.in
docutils==0.16            # via sphinx
feedparser==6.0.2         # via -r requirements.in
filelock==3.0.12          # via pytest-mypy, tox, virtualenv
flake8==3.8.4             # via -r requirements.in, pytest-flake8
idna==2.10                # via requests
imagesize==1.2.0          # via sphinx
iniconfig==1.1.1          # via pytest
isort==5.6.4              # via -r requirements.in, pytest-isort
jinja2==2.11.2            # via sphinx
livereload==2.6.3         # via sphinx-autobuild
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
model-bakery==1.2.1       # via -r requirements.in
mypy-extensions==0.4.3    # via black, mypy
mypy==0.790               # via pytest-mypy
packaging==20.4           # via bleach, pytest, sphinx, tox
pathspec==0.8.1           # via black
pip-tools==5.3.1          # via -r requirements.in
pluggy==0.13.1            # via pytest, tox
py==1.9.0                 # via pytest, tox
pycodestyle==2.6.0        # via flake8
pyflakes==2.2.0           # via flake8
pygments==2.7.2           # via sphinx
pyparsing==2.4.7          # via packaging
pytest-black==0.3.12      # via -r requirements.in
pytest-cov==2.10.1        # via -r requirements.in
pytest-django==4.1.0      # via -r requirements.in
pytest-flake8==1.0.6      # via -r requirements.in
pytest-isort==1.2.0       # via -r requirements.in
pytest-mypy==0.8.0        # via -r requirements.in
pytest==6.1.2             # via -r requirements.in, pytest-black, pytest-cov, pytest-django, pytest-flake8, pytest-mypy
pytz==2020.4              # via babel, django
regex==2020.11.13         # via black
requests==2.25.0          # via sphinx
sgmllib3k==1.0.0          # via feedparser
six==1.15.0               # via bleach, livereload, packaging, pip-tools, tox, virtualenv
snowballstemmer==2.0.0    # via sphinx
sphinx-autobuild==2020.9.1  # via -r requirements.in
sphinx==3.3.1             # via -r requirements.in, sphinx-autobuild
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
sqlparse==0.4.1           # via django
toml==0.10.2              # via black, pytest, pytest-black, tox
tornado==6.1              # via livereload
tox==3.20.1               # via -r requirements.in
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.3  # via black, mypy
urllib3==1.26.2           # via requests
virtualenv==20.1.0        # via tox
webencodings==0.5.1       # via bleach

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
